front-end: Produce data for live demo (#16)
To achieve this, create a dedicated populatedb-demo command, with
customised values. Also, update DB exports accordingly.

Charge_session fields `energy_delivered` and `total_cost` are converted
to floats, for more realistic data. Migrations are updated accordingly.

+++ back-end/api/models/charge_session.py
@@ -20,8 +20,8 @@ class ChargeSession(models.Model):
             on_delete=models.SET_NULL)
     pricing_id = models.ForeignKey(Pricing, null=True,
             on_delete=models.SET_NULL)
-    energy_delivered = models.PositiveIntegerField()
-    total_cost = models.PositiveIntegerField()
+    energy_delivered = models.FloatField()
+    total_cost = models.FloatField()
     start = models.DateTimeField()
     end = models.DateTimeField()
     payment = models.CharField(
